pyproject.toml
# Build System
[build-system]
requires = ["setuptools"]

# Project Metadata
[project]
name = "checkmark"
version = "0.1.0"
description = "A tool to automatically create and evaluate assessments."
readme = "README.md"
requires-python = ">=3.8"
license = { file = "LICENSE" }
authors = [
    { name = "Daniel Mizsak", email = "info@pythonvilag.hu" },
    { name = "Samu Pecsenye", email = "samu.pecsenye@gmail.com" },
]
keywords = ["assessment", "evaluation", "automatic", "checkmark"]
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Education",
]
dependencies = [
    # Generator
    "cryptography==40.0.2",
    "fpdf2==2.5.7",
    "matplotlib==3.6.3",
    "openpyxl==3.0.10",
    "pandas==1.5.1",
    "qrcode==7.3.1",
    "requests==2.28.1",
    "ttkbootstrap==1.10.1",
    # Evaluator
    "opencv-python==4.6.0.66",
    "pillow_heif==0.11.0",
    "pyzbar==0.1.9",
    #Server
    "Flask==2.3.2",
]
scripts.checkmark = "checkmark.main:main"

[project.optional-dependencies]
dev = [
    "black",
    "coverage",
    "mypy",
    "pre-commit",
    "pytest-cov",
    "sphinx",
    "pytest",
    "ruff",
]

[project.urls]
Repository = "https://github.com/daniel-mizsak/checkmark"
Homepage = "https://pythonvilag.hu/pv/projects/checkmark"

# Tools
# Setup Tools
[tool.setuptools.packages.find]
where = ["src"]

# Black
[tool.black]
line-length = 100

# MyPy
[tool.mypy]
mypy_path = ["src", "tests"]
check_untyped_defs = true
disallow_any_generics = true
disallow_incomplete_defs = true
ignore_missing_imports = true
no_implicit_optional = true
show_error_codes = true
warn_redundant_casts = true
warn_return_any = true
warn_unreachable = true
warn_unused_configs = true
warn_unused_ignores = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

# Pytest
[tool.pytest.ini_options]
addopts = "--cov=checkmark"
testpaths = ["tests"]
filterwarnings = ["ignore::DeprecationWarning"]

# Ruff
[tool.ruff]
line-length = 100


# Tox
# TODO: Add coverage and black
[tool.tox]
legacy_tox_ini = """
    [tox]
    envlist = py38,py311,mypy
    skip-missing-interpreters = true

    [gh]
    python =
        3.8 = py38
        3.11 = py311, mypy

    [testenv]
    setenv =
        PYTHONPATH = {toxinidir}/src
    deps =
        pytest
        pytest-cov
    commands = pytest {posargs:tests}

    [testenv:mypy]
    basepython = python3.11
    deps = mypy
    commands = mypy src tests
"""
